requirements.txt
# API 服务
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 项目已有依赖
redis>=5.0.0
chromadb>=0.4.0
openai>=1.0.0
sentence-transformers>=2.2.0
pyyaml>=6.0
neo4j>=5.0.0
httpx>=0.24.0
spacy>=3.5.0
datasets>=2.14.0
langdetect>=1.0.0
